# Rc containers were removed from LocalChannel implementation.