--- a/glommio/src/channels/mod.rs
+++ b/glommio/src/channels/mod.rs
@@ -20,8 +20,8 @@ mod spsc_queue;
 /// The communication between sender and receiver is broken when one of them goes
 /// out of scope. They however behave differently:
 ///
-/// * The [`LocalReceiver`] never sees an error, as it is implemented as a stream
-///   interface compatible with [`StreamExt`]. When the sender is no longer available the receiver's call to [`next`] will return [`None`].
+/// * The [`LocalReceiver`] never sees an error, as it provides a stream
+///   interface compatible with [`StreamExt`]. When the sender is no longer available the stream's call to [`next`] will return [`None`].
 /// * The [`LocalSender`] will return a [`ChannelError`] encapsulating a [`BrokenPipe`] error if it tries to [`send`] into
 ///   a channel that no longer has a receiver.
 ///
@@ -40,7 +40,7 @@ mod spsc_queue;
 ///
 ///     let (sender, mut receiver) = local_channel::new_unbounded();
 ///     let h = Local::local_into(async move {
-///         assert_eq!(receiver.next().await.unwrap(), 0);
+///         assert_eq!(receiver.stream().next().await.unwrap(), 0);
 ///     }, task_queue).unwrap().detach();
 ///     sender.try_send(0);
 ///     drop(sender);