### What does this PR do?

Rc containers were removed from LocalChannel implementation.

Because Rc is a boxed type, Rc containers which were used inside of 
LocalChannel implementation of the waiting queue 
were replaced by NonNull-s of Pinned waiting nodes.
To ensure safety the following changes were done:
1. Futures lifetimes were restricted by lifetime of the LocalChannel.
 That is quite a minor restriction because 
that is typically how sync primitives are used anyway.
2. All waiting nodes were made Pinned.
3. On Drop of the LocalChannel assertion that waiting queues are empty 
was added.
4. On Drop Futures remove itself from waiting queue.
5. LocalChannel does not implement Stream itself but returns related Stream 
instance through LocalReceiver#stream() method. 
That makes usage less convenient but a) It is more canonical because single 
Stream should work with single Task instance the same as Future. 
b) It makes implementation simpler.
6. On Drop of stream associated waiting node is removed from the waiting queue. 

Results of benchmark

before 

cost of sending contended local channel: 137ns
cost of sending uncontended local channel: 40ns

after 

cost of sending contended local channel: 138ns 0%
cost of sending uncontended local channel: 17ns  -58%

Code is checked by ASAN.


### Motivation

Decrease heap allocations during the scheduling of tasks

### Related issues

Issue #195

### Checklist

[] I have added unit tests to the code I am submitting
[] My unit tests cover both failure and success scenarios
[] If applicable, I have discussed my architecture
[x] The new code I am adding is formatted using `rustfmt`