@laa 2020-12-18
null pointer safety checks were added

Waiter node references converted into the Pin-ed whenever possible.