@glommer 2020-12-18
@laa 

clippy from beta is complaining about this now:

```
warning: explicit lifetimes given in parameter types where they could be elided (or replaced with `'_` if needed by type declaration)
   --> glommio/src/channels/local_channel.rs:732:5
    |
732 |     pub fn stream<'a>(&'a self) -> impl Stream<Item = T> + 'a {
    |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    |
    = note: `#[warn(clippy::needless_lifetimes)]` on by default
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#needless_lifetimes
```

I am already using the beta stream locally, but not mandating in CI yet.