@glommer 2020-12-18
this is FYI. It's a trivial fix, so I will push a fix.